+++ pkg/metadata/annotations.go
@@ -17,6 +17,7 @@ package metadata
 import (
 	"strings"
 
+	"github.com/pulumi/pulumi/pkg/resource"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
@@ -48,9 +49,17 @@ func IsInternalAnnotation(key string) bool {
 // SetAnnotation sets the specified key, value annotation on the provided Unstructured object.
 func SetAnnotation(obj *unstructured.Unstructured, key, value string) {
 	// Note: Cannot use obj.GetAnnotations() here because it doesn't properly handle computed values from preview.
+	// during preview, our strategy for if metdata or annotations end up being computed values, is to just not
+	// apply an annotiation (since there's no way to insert data into the computed object)
 	metadataRaw := obj.Object["metadata"]
+	if isComputedValue(metadataRaw) {
+		return
+	}
 	metadata := metadataRaw.(map[string]interface{})
 	annotationsRaw, ok := metadata["annotations"]
+	if isComputedValue(annotationsRaw) {
+		return
+	}
 	var annotations map[string]interface{}
 	if !ok {
 		annotations = make(map[string]interface{})
@@ -78,4 +87,9 @@ func IsAnnotationTrue(obj *unstructured.Unstructured, key string) bool {
 func GetAnnotationValue(obj *unstructured.Unstructured, key string) string {
 	annotations := obj.GetAnnotations()
 	return annotations[key]
+}
+
+func isComputedValue(v interface{}) bool {
+	_, isComputed := v.(resource.Computed)
+	return isComputed
 }